Drag based on the picked element. Fixes issue 17. Anyway dragging doesn't change the selection (this is a feature!).
Update issue 17
The selection is not affected by dragging. This allows you to drag a node which
is not selected without having your selection cleared. The motivation is that
clearing the selection is a simple matter of clicking on the background, while
to re-create it you would have to re-select all the elements.